Refactor TelemetryFlattener to TelemetryLabeler, add OptionalTelemetryLabeler
